usecase 작업 완료

+++ apps/src/main/kotlin/io/dustin/apps/board/domain/qna/question/repository/custom/impl/CustomQuestionRepositoryImpl.kt
@@ -43,7 +43,7 @@ class CustomQuestionRepositoryImpl(private val query: JPAQueryFactory) : CustomQ
         return jPAQuery.fetchOne()
     }
 
-    override fun getQuestionList(userId: Long, nextId: Long?, size: Int): List<QuestionDto> {
+    override fun getQuestionList(userId: Long, nextId: Long?, size: Long): List<QuestionDto> {
         val question = QQuestion.question
         val answer = QAnswer.answer
